New Fluid2EXR to replace Smoke2EXR
The old Smoke2EXR used the Point Cache file to extract smoke information into a set of EXR images. This is slow and relies on files being in Point Cache format. The new add-on uses the data available in the Fluid modifier directly - which makes it compatible with the Mantaflow fluid in Blender 2.82+